requirements.txt
Django>=4.3
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
gunicorn